% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularity.R
\name{circularity}
\alias{circularity}
\alias{perimeter}
\title{Improved Circularity and Perimeter Calculation}
\usage{
circularity(x)

perimeter(x)
}
\arguments{
\item{x}{An \code{Image} object mask with a single integer or binary
object to be analyzed.}
}
\value{
Circularity as a value between 0 (line) and 1 (circle)

The perimeter (in pixels) of a binary object
}
\description{
Improved Circularity and Perimeter Calculation
}
\details{
These functions provide more precise measures of the circularity and
perimeter of an integer or binary object than the algorithm in
 \code{\link[EBImage]{computeFeatures}}. Values are computed
by the C++ code (\code{circularity_cpp} and \code{perimeter_cpp}).
Because of digitization, the circularity can be greater
than 1. This function truncates values greater than 1. Also,
line-like objects with no interior pixels return an inappropriately
high value for the circularity. The perimeter for such objects is
calculated differently by treating each pixel as a unit square and
counting their exposed sides.
}
