% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{dp}
\alias{dp}
\title{Plot Device Wrapper}
\usage{
dp(path, save = SaveImage, width, dev, res = 300, units = "in", ...)
}
\arguments{
\item{path}{file path to save the device window. If missing, 
\code{\link{file.choose}} will be called to select the file}

\item{save}{\code{logical} value to save image. If missing, the global 
environment will be searched for a \code{logical} value named 'SaveImage'}

\item{width}{device width in units specified by \code{units}. If missing
the width of the current device will be used}

\item{dev}{suitable printing device, defaults to \code{\link{png}} but 
can be one of \code{\link{bmp}}, \code{\link{jpeg}}, or \code{\link{tiff}}. 
If not specified, the device will be determined by the path extension}

\item{res}{resolution in pixels per \code{units}, default of 300}

\item{units}{the units in which \code{width} is given. Defaults to "in" 
(inches) rather than the normal default of \code{px} (pixels)}

\item{...}{additional paramters passed to \code{\link{dev.print}}}
}
\value{
The basename of the saved file will be returned or \code{NULL}
if nothing was saved.
}
\description{
This is a wrapper to plot the current device to a file.
}
\details{

}
\seealso{
\code{\link{dev.print}} that is called by this function
}
