% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBImageExtra-package.R
\docType{package}
\name{EBImageExtra-package}
\alias{EBImageExtra}
\alias{EBImageExtra-package}
\title{EBImageExtra: Helper Functions for EBImage}
\description{
Helper functions extending the functionality of EBImage.
}
\author{
\strong{Maintainer}: David Ornelles \email{david.ornelles@wfusm.edu}

}
\keyword{internal}
