% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameROI.R
\name{frameROI}
\alias{frameROI}
\title{Draw a Frame/Border around an Image}
\usage{
frameROI(img, lwd = 2, col = "white", sides = 1:4)
}
\arguments{
\item{img}{An \code{Image} object}

\item{lwd}{'line' width of the border in pixels; note that this is
\strong{not} the standard definition of \code{lwd}}

\item{col}{border color of the rectangle}

\item{sides}{which side(s) are to include the border (1=bottom, 2=left,
3=top, 4=right)}
}
\value{
Modified image with border
}
\description{
Draw a rectangular frame about the border of an image along
the indicated sides
}
\details{
A rectangular border of \code{lwd} pixels will drawn around the
image along the sides identified in \code{sides}. The added
border will be drawn toward the interior of the image. This
differs from using \code{link{rect}} with \code{lwd} used in an atypical
manner. Here it defines the width of the border in pixels. This also differs
from the base functions by directly changing pixels in the image.
}
\examples{
	lighthouse <- readImage(system.file("inst", "extdata", "lighthouse.jpg", package="EBImageExtra"))
plot(frameROI(lighthouse), col = "black") # with defaults
 plot(frameROI(lighthouse, lwd = 36, col = "yellow", sides = c(1,2)))

}
\seealso{
\code{\link{getROI}} to get a region of interest from an image;
\code{\link{putROI}} to place an ROI with scaling;
\code{\link{drawROI}} to draw a frame \emph{within} an image;
\code{\link{insertROI}} as a convenience function that
  combines calls to \code{getROI}, \code{putROI}
  and \code{drawROI} to place a framed inset in an image.
}
