% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFrames.R
\name{getFrames}
\alias{getFrames}
\title{Corrected getFrames function from EBImage}
\usage{
getFrames(y, i, type = c("render", "total"))
}
\arguments{
\item{y}{An \code{Image} object or array.}

\item{i}{Number of frames as a single number of vector of numbers. If
missing, all frames are returned.}

\item{type}{A character string with \code{total} or \code{render}. The
\emph{new} default is \code{render}.}
}
\value{
Returns i-th frame(s) containing the image \code{y}. If \code{type} is
\code{total}, the total number of frames is independent of the color mode
and equals the product of all dimensions except the first two.
}
\description{
This corrects the frame argument error in the EBImage code and sets the 
default from "total" to "render".
}
\seealso{
\code{\link[EBImage]{getFrame}}
}
