% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getROI.R
\name{getROI}
\alias{getROI}
\title{Get Region of Interest}
\usage{
getROI(img, x, y, x2, y2, w, h, asCorner = FALSE,
  which.corner = c("bottomleft", "topleft", "bottomright", "topright"),
  pch = 3, col = "magenta", border = col, lwd = 2, markup = TRUE)
}
\arguments{
\item{img}{An \code{Image} object}

\item{x, y}{\code{x,y} coordinates of one or both corners of
rectangular selection \emph{or} the location of the center of the
rectangular selection \emph{or} a list specifying
the corners of the selection (see details)}

\item{x2, y2}{optional second pair of x and y coordinates when
needed to specify the other corner of the rectangular selection}

\item{w, h}{optional width and height of the rectangular selection;
required if \code{x2,y2} are missing}

\item{asCorner}{\code{logical} value to use the
point \code{x,y} as the corner of the selection or as the center of
the selection}

\item{which.corner}{identifies the
corner of the rectangle specified by \code{x,y};
applies only if \code{asCorner = TRUE}}

\item{pch}{plotting character used by \code{locator}, default of 3 (cross)
when interacting with the image}

\item{col}{color for plotting character used by \code{locator}, default of
\code{"magenta"}; use \code{NA} for no plotting character}

\item{border}{border color of rectangle if \code{markup = TRUE}; if not
specified, the value for \code{col} is used}

\item{lwd}{line width of rectangle if \code{markup = TRUE}}

\item{markup}{\code{logical} value (default of \code{TRUE}) indicating
that the image should be redrawn with the selection outline drawn}
}
\value{
A list of two named objects with \code{"image"} as the extracted region of
interest as an \code{Image} and \code{"roi"} as the \code{x,y} coordinates
for the region of interest as a list.
}
\description{
Get a rectangular region of interest from an image.
}
\details{

}
\section{Selecting the region of interest}{

This permits the interactive or programmatic definition of a rectangular
region of interest in an image by either specifying
coordinates that define the rectangle or by invoking \code{\link{locator}}
to allow the user to select the region of interest. The region of interest
will be trimmed to the dimensions allowed by the original image.
Options allow specifying the rectangle either by
the center or corner(s) as describe below. The first two options require
no interaction with the user and only produce an image if
\code{markup = TRUE}. Options 3 and 4 below require interaction with the
user. In all cases, the selection coordinates are adjusted to conform to
the dimensions of the image. 

\enumerate{
  \item{\strong{List}.} If \code{x} is a \code{list} of length 2, it
    is assumed to hold opposite corners of the rectangular selection. This
    is the typical result of a call to \code{locator(2)} after having
    plotted the image.
  \item{\strong{Two Points}.} If values are provided for each of \code{x,y} and
    \code{x2,y2}, these are treated as opposite corners of
    the rectangular selection.
  \item{\strong{One Point} (with width and height).} If only two values are
    provided for \code{x,y} and \code{x2,y2}, they will be assigned to
    \code{x,y}. In this case, values for \code{w,h} must be provided as
    named arguments for the width and height of the rectangular selection.
    The point \code{x,y} is interpreted as \emph{either} the center
    (\code{asCorner=FALSE})
    \emph{or} the corner of the selection (\code{asCorner=TRUE}). If
    \code{asCorner = TRUE}, the position of the corner is determined
    by the argument \code{which.corner} which can be one of \code{"bottomleft", 
    "topleft", "bottomright",} or \code{"topright"}.
  \item{\strong{No Points} (choose opposite corners).} If all of \code{x,y, x2,y2, w,h}
    are missing, \code{\link{locator}} will be called to let the user to
    select two points that define opposite corners of the rectangular selection.
}
}

\section{Return value}{

The returned value is a list of length 2 with the image (\code{"image"})
and the region of interest (\code{"roi"}) as a list of two \code{x,y} points
defining the region of interest.

The returned \code{"roi"} component can be used to extract the equivalent
region of interest from an identically sized image or frame
with code such as:
\preformatted{
   pp <- getROI(img)$roi
   pp <- lapply(pp, function(v) seq.int(v[1], v[2]))
   idx <- lapply(dim(img), seq.int)
   idx[1:2] <- pp
   cropped <- do.call("[", c(list(img), idx))
}
}

\examples{
# Example using fixed width and height to retrieve image
  lighthouse <- readImage(system.file("inst", "extdata", "lighthouse.jpg", package="EBImageExtra"))
# Get region of interest
  ans <- getROI(lighthouse, 515, 275, w = 200, h = 300)
  plot(ans$image)
 print(ans$roi)
# Get region of interest as a pair of points, anchored by top left corner
  ans <- getROI(lighthouse, 515, 275, w = 200, h = 300, asCorner = TRUE,
     which.corner = "topleft")
  print(ans$roi)

}
\seealso{
addInset; other stuff
}
