% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpf.R
\name{hpf}
\alias{hpf}
\title{High Pass Filter}
\usage{
hpf(x, sigma = 21, min.value = 2^-12)
}
\arguments{
\item{x}{An \code{Image} object or array}

\item{sigma}{A numeric passed to \code{gblur} as the standard deviation
of the Gaussian filter used for initial blurring. This should be
approximately the minimum size of the feature to keep}

\item{min.value}{Numeric value for the minimum value returned by this
function}
}
\value{
An image object of the same size with slowly varying signals removed.
}
\description{
Attenuate slowly varying signals in an \code{Image} object
}
\details{
This function subtracts the blurred image from the original image and
adds the mean value of the original image. The resulting pixels are
trimmed so that the lowest value is \code{min.value}.
}
