% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roi-methods.R
\name{loc}
\alias{loc}
\title{Accessor functions for class "Roi"}
\usage{
loc(x, ...)
}
\arguments{
\item{x}{object of class \code{roi} created by \code{Roi}}
}
\value{
list of original location in the slot \code{loc}
}
\description{
Presently, only one to extract the \code{loc} slot
}
