% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niq.R
\name{niq}
\alias{niq}
\title{Normalize Each Frame of an Image to an Upper "Quantile"}
\usage{
niq(x, quant = 0.998, ft = c(0, 1))
}
\arguments{
\item{x}{An \code{Image} object or array or a list of \code{Image} objects}

\item{quant}{A numeric value between 0 and 1 specifying the upper
quantile for the normalization.}

\item{ft}{A numeric vector of 2 values or \code{list} of vectors to be the
target minimum and maximum intensity values after normalization.}
}
\value{
An \code{Image} object or array or a \code{list} of such objects containing
the transformed object(s).
}
\description{
This function normalizes each frame of an image to the range specified
specified by \code{quant} (default quantile of 0.998.)
}
\details{
See \code{\link[EBImage]{normalize}} for details as well as \code{ni}.
}
