% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeROI.R
\name{placeROI}
\alias{placeROI}
\title{Place Image (ROI) into Another Image}
\usage{
placeROI(roi, img, position, frac = NULL, mag = NULL,
  frac.default = 1/3, lwd = 2, col = "white", showImage = TRUE)
}
\arguments{
\item{roi}{the smaller \code{Image} object to be placed in
\code{img} \emph{or} the filename (as \code{character} vector) of
  such an image}

\item{img}{the larger \code{Image} object to receive the roi}

\item{position}{a character string or integer in 1 to 9 indicating
the position of the inset; If \code{missing}, \code{\link{locator}}
will be used to determine where to place the inset; see Details}

\item{frac}{\code{numeric} value between 0 and 1 for the fractional
width occupied by the inset; the default values of \code{NULL} uses the 
value in \code{frac.default}}

\item{mag}{optional magnification factor for inset; if \code{NULL}
(default), it will be determined by \code{frac}}

\item{frac.default}{default fractional width for the inset (\code{1/3})}

\item{lwd}{'line' width of the inset border in pixels; note that this is
\strong{not} the standard definition of \code{lwd}}

\item{col}{color of the inset border}

\item{showImage}{\code{logical} value indicated whether to plot the 
image with inset; default (\code{TRUE})}
}
\value{
An \code{Image} of the same dimensions as the second argument
(\code{img}) with \code{roi} inserted after appropriate scaling.
}
\description{
Place a scaled image into another image as an inset
}
\details{
The image to be placed as an inset (\code{roi}) will be scaled 
and placed in \code{img} at the location specified by \code{position}. 
This argument can be one of "topleft", "top", "topright", "left", 
"center", "right", "bottomleft", "bottom", and "bottomright". The code 
will also accept the integers 1 to 9 corresponding to these
respective positions. If \code{position} is missing,
\code{\link{locator}} will be called to determine where (among the
nine choices) to place the inset. 

The image will be scaled according to either \code{frac} or \code{mag} 
where \code{frac} indicates the fractional width of the image to be 
occupied by the scaled inset. Alternatively, the magnification can be 
specified with \code{mag}, typically a number greater than 1. Values that
 create an inset greater than the dimensions of the \code{img} argument 
will cause an error. 

If \code{showImage = TRUE}, the image with inset will be plotted. In all
cases, the modified image will be invisibly returned.
}
\examples{
# To be done...

}
\seealso{
\code{\link{getROI}}; \code{\link{frameROI}}
}
