% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleBar2.R
\name{scaleBar2}
\alias{scaleBar2}
\title{Add Scale Bar to Image}
\usage{
scaleBar2(
  x,
  y = NULL,
  distance,
  scale,
  units = "µm",
  col = "white",
  col.text,
  cex = 3/4,
  adj = c(0.5, -0.5),
  col.line,
  lwd = 1,
  lend = 1,
  xpd = NA,
  ...
)
}
\arguments{
\item{x, y}{Central coordinates of the scale bar in pixels}

\item{distance}{Distance represented by the scale bar in microns}

\item{scale}{Image scale as µm per pixel}

\item{units}{Unit of measure, default of "µm". Use \code{NA} to exclude
the label}

\item{col}{Default color (white) for \code{col.text} and \code{col.line}}

\item{col.text, cex, adj, xpd, ...}{Parameters passed to the \code{\link{text}}
function used to add the label}

\item{col.line, lwd, lend}{Parameters passed to the \code{\link{lines}}
function used to draw the horizontal scale bar}
}
\value{
This function is called for side effects but the location is returned
invisibly as a list.
}
\description{
Place a single horizontal scale bar of length \code{distance} as per
scale factor \code{scale} with unit of measure \code{units}.
}
\details{
This is a utility to place a labeled scale bar on a raster image. 
A default label will include the distance in µm if \code{unit} is not
\code{NA}. The label is centered at the position specified by \code{x,y}
according to \code{adj}. The default value for \code{adj} of
\code{c(0.5,-0.5)} centers the label and places it 0.5 character widths
above \code{x,y}.
 
The \code{x,y} coordinates can be passed in a plotting structure
appropriate for \code{\link{xy.coords}}. This function attempts to parse
\code{distance} and \code{scale} from the remaining arguments if they are
not named and if \code{x} is a list with components named "x" and "y". This 
allows one to interactively place a scale bar with a function
call such as in the following example. In this example, \code{scale} is
the scale factor as "µm per pixel."

This call will place a labeled 25 µm scale bar on the image as per the 
value in \code{scale}
\preformatted{
> scaleBar(locator(), 25, scale)
}

Note that this is \strong{not} a vectorized function. Only the last
value of \code{(x, y)} will be used to place the scale bar. This allows
multiple clicks to \code{locator()} to get the position just right
with only the last click being used.

Micrometers-per-pixels calibration values for the Nikon TE300 at 1x1
binning with the QImaging camera are:
\preformatted{
Objective   scale factor
 20X ELWD    0.3228
 40X ELWD    0.1624
 60X ELWD    0.1062
 60X Apo     0.1064
100X Apo     0.0640
  4X Apo     1.613 (estimated)
}
}
