% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobel.R
\name{sobel}
\alias{sobel}
\title{Sobel Filter for Edge Detection}
\usage{
sobel(x)
}
\arguments{
\item{x}{An \code{Image} object or array}
}
\value{
An image object of the same size with edges highlighted.
}
\description{
Detect edges in an \code{Image} object
}
\details{
The following matrix and its transpose are sequentially applied in
\code{\link[EBImage]{filter2}} to identify edges in the image as described
in this \href{https://en.wikipedia.org/wiki/Sobel_operator}{Wikipedia} page. 

\preformatted{
     [,1] [,2] [,3]
[1,]    1    0   -1
[2,]    2    0   -2
[3,]    1    0   -1
}
}
