% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tStack.R
\name{tStack}
\alias{tStack}
\title{Transpose an Array of Images}
\usage{
tStack(x, n = 2)
}
\arguments{
\item{n}{the number of adjacent images (default of 2)}

\item{stk}{array of 3- or 4-dimensions \code{Image}
objects}
}
\value{
array of \code{Image} objects with the final dimension permuted
}
\description{
Transpose the final dimension of a 3- or 4-dimensional array (stack)
}
\details{
This function will "transpose" an
\code{Image} object with 3 or 4 dimensions. The motivation behind this
function is simplify the task of reordering an array generated by
sequentially combining several image groups so that related images are
adjacent.

As a typical example, three different images (DAPI, FITC, and Cy3) are
taken of four separate fields of cells for 12 images. The images are
separated, colorized, and then combined with \code{combine(...)} or
code{abind(..., along=4)}. Plotting as \code{plot(..., all= TRUE, nx =
4)} shows the same field arranged vertically. Transposing this image
allows it to be presented horizontally.
}
