% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI-methods.R
\name{as.Roi}
\alias{as.Roi}
\alias{as.ROI}
\title{Accessory functions for class "Roi"}
\usage{
as.Roi(img, ...)

as.ROI(img, ...)
}
\arguments{
\item{img}{object of class \code{Image}}

\item{resize}{re-assign \code{"loc" slot} if \code{TRUE}}
}
\value{
\code{Roi} object of the same dimensions as the
}
\description{
Convert an \code{Image} object into an \code{Roi} object
}
\details{
This adds the \code{class} "Roi" to \code{Image} objects and places the
dimensions of the image in the \code{slot} named "\code{loc}."
\code{Roi} objects passed as an argument will be considered an error
\strong{unless} \code{resize = TRUE}.
}
\examples{

# Region of interest from sample image in EBImage package
  birds <- readImage(system.file("images", "sample-color.png", package="EBImage"))
  roi <- getROI(birds, 160, 255, w = 200, h = 240)

# Examine effect on Image object
  attr(birds, "loc") # nothing here
  attr(as.Roi(birds), "loc") # returned an Roi

# Examine effect on 'Roi'
  pp <- attr(roi, "loc") # original
  roi2 <- resize(roi, w = 400)
  pp2 <- attr(roi2, "loc") 
  identical(pp, pp2) # re-sizing the image has no effect on 'loc'

# Compare with resize = TRUE
  attr(roi, "loc") 
  attr(as.Roi(roi, resize = TRUE), "loc") 

}
