% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roi-methods.R
\name{as.Roi}
\alias{as.Roi}
\title{Accessory functions for class "Roi"}
\usage{
as.Roi(x)
}
\arguments{
\item{x}{object of class \code{Image}}
}
\value{
\code{Roi} object of the same dimensions as the 
  argument but with a \code{slot} named \code{loc}
  holding the dimension of the image as the region
  of interest. \strong{Note that this will reset
  the \code{loc} slot an \code{Roi} object.}
}
\description{
Presently only one function to convert an \code{Image} object
into an \code{Roi} object
}
