% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularity.R
\name{circularity}
\alias{circularity}
\alias{perimeter}
\title{Improved Circularity and Perimeter Calculation}
\usage{
circularity(x)

perimeter(x)
}
\arguments{
\item{x}{An \code{Image} object mask with a single object to be analyzed.}
}
\value{
Circularity as a value between 0 (line) and 1 (circle)

The perimeter (in pixels) of a binary object
}
\description{
Improved Circularity and Perimeter Calculation
}
\details{
At the cost of increased computing time, these functions provide more precise
measures of the circularity and perimeter of a binary object than the
algorithm in \code{\link[EBImage]{computeFeatures}}.
}
