% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{crop}
\alias{crop}
\title{Crop an Image to Visible Pixels}
\usage{
crop(img, border = 1, fill = 0)
}
\arguments{
\item{img}{A grayscale or binary \code{Image} object of two dimensions}

\item{border}{Single integer value of border to add to cropped image}

\item{bg}{Background value for border}
}
\value{
A cropped image with the specified border
}
\description{
Crop a grayscale or binary \code{Image} object to non-zero values plus a
given border
}
\details{
Non-zero pixels on the edge of the image are removed, leaving a border of
\code{border} pixels of \code{bg} color.
}
