% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawROI.R
\name{drawROI}
\alias{drawROI}
\title{Draw a Rectangular Frame within an Image}
\usage{
drawROI(img, x, y, x2, y2, w, h, lwd = 2, col = "white",
  asCorner = FALSE, which.corner = c("bottomleft", "topleft",
  "bottomright", "topright"), sides = 1:4, pch = 3, col.pch = col,
  showImage = TRUE)
}
\arguments{
\item{img}{An \code{Image} object}

\item{x, y}{\code{x,y} coordinates of one or both corners of
rectangular selection \emph{or} the location of the center of the
rectangular selection \emph{or} a list specifying
the corners of the selection (see details)}

\item{x2, y2}{optional second pair of x and y coordinates when
needed to specify the other corner of the rectangular selection}

\item{w, h}{optional width and height of the rectangular selection;
required if \code{x2,y2} are missing}

\item{lwd}{'line' width of the border in pixels; note that this is
\strong{not} the standard definition of \code{lwd}}

\item{col}{border color of the rectangle}

\item{asCorner}{\code{logical} value to use the
point \code{x,y} as the corner of the selection or as the center of
the selection}

\item{which.corner}{identifies the
corner of the rectangle specified by \code{x,y};
applies only if \code{asCorner = TRUE}}

\item{sides}{which side(s) are to include the border (1=bottom, 2=left,
3=top, 4=right)}

\item{pch}{plotting character used by \code{link{locator}} when interacting
with the image}

\item{col.pch}{color for plotting character used by \code{link{locator}}}

\item{showImage}{replot the image with border if \code{TRUE}}
}
\value{
Modified image with region of interest outlined.
}
\description{
Draw a rectangular frame surrounding the region of
interest in an image.
}
\details{
A rectangular border of \code{lwd} pixels will drawn around the
region of interest along the sides identified in \code{sides}. The added
border will be drawn toward the interior of the region of interest. This
differs from using \code{link{rect}} with \code{lwd} used in an atypical
manner. Here it defines the width of the border in pixels. This also differs
from the base functions by directly changing pixels in the image.

This region of interest can be specified with arguments or interactively. 
Options allow specifying the rectangle either by
the center or corner(s) as describe below. The first two options require
no interaction with the user and will display the revised image
only if \code{showImage = TRUE}. Options 3 and 4 below require interaction
with the user. In all cases, the selected coordinates are adjusted to conform
to the dimensions of the image, which can override values in \code{w,h}.  

\enumerate{
  \item{\strong{List}.} If \code{x} is a \code{list} of length 2, it
    is assumed to hold opposite corners of the rectangular selection. This
    could be the \code{loc} slot from a \code{Roi} object.
  \item{\strong{Two Points}.} If values are provided for each of \code{x,y} and
    \code{x2,y2}, these are treated as opposite corners of
    the rectangular selection.
  \item{\strong{One Point} (with width and height).} If only two values are
    provided for \code{x,y} and \code{x2,y2}, they will be assigned to
    \code{x,y}. In this case, values for \code{w,h} must be provided as
    named arguments for the width and height of the rectangular selection.
    The point \code{x,y} is interpreted as \emph{either} the center
    (\code{asCorner=FALSE})
    \emph{or} the corner of the selection (\code{asCorner=TRUE}). If
    \code{asCorner = TRUE}, the position of the corner is determined
    by the argument \code{which.corner} which can be one of \code{"bottomleft", 
    "topleft", "bottomright",} or \code{"topright"}.
  \item{\strong{No Points} (choose opposite corners).} If all of \code{x,y, x2,y2, w,h}
    are missing, \code{\link{locator}} will be called to let the user to
    select two points that define opposite corners of the rectangular selection.
}
}
\seealso{
\code{\link{getROI}} to get a region of interest from an image;
\code{\link{putROI}} to place an ROI with scaling;
\code{\link{frameROI}} to draw a frame \emph{about} an image;
\code{\link{insertROI}} as a convenience function that
  combines calls to \code{getROI}, \code{putROI}
  and \code{drawROI} to place a framed inset in an image.
}
