% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawROI.R
\name{drawROI}
\alias{drawROI}
\title{Draw a Rectangular Frame within an Image}
\usage{
drawROI(img, x, y, x2, y2, w, h, show, lwd = 2, col = "white",
  asCorner = FALSE, which.corner = c("bottomleft", "topleft",
  "topright", "bottomright"), sides = 1:4, pch = 3, col.pch = col)
}
\arguments{
\item{img}{An \code{Image} object}

\item{x, y}{\code{x,y} coordinates of one or both corners of
rectangular selection \emph{or} the location of the center of the
rectangular selection \emph{or} a list specifying
the corners of the selection \emph{or} an \code{Roi} object
with a \code{slot} named "\code{loc}" (See Details)}

\item{x2, y2}{optional second pair of x and y coordinates when
needed to specify the other corner of the rectangular selection}

\item{w, h}{optional width and height of the rectangular selection;
required if \code{x2,y2} are missing}

\item{show}{replot the image after adding the border if \code{TRUE}}

\item{lwd}{line width of the border in pixels; note that this is
\strong{not} the standard definition of \code{lwd}}

\item{col}{border color of the rectangle}

\item{asCorner}{\code{logical} value to use the point \code{x,y} 
as the corner of the selection instead of the center of
the selection}

\item{which.corner}{identifies the corner of the rectangle specified
by \code{x,y}; applies only if \code{asCorner = TRUE}}

\item{sides}{which side(s) are to include the border (1=bottom, 2=left,
3=top, 4=right)}

\item{pch}{plotting character used by \code{\link{locator}} when interacting
with the image}

\item{col.pch}{color for plotting character used by \code{\link{locator}}}
}
\value{
Modified image with region of interest outlined.
}
\description{
Draw a rectangular frame surrounding the region of
interest in an image or the entire image.
}
\details{
A rectangular border of \code{lwd} pixels will drawn around the
region of interest along the sides specified by \code{sides}. The added
border will be drawn toward the interior of the region of interest to
preserve the dimensions of the modified image. This function
differs from \code{\link{rect}} by using \code{lwd} in an atypical
manner--here \code{lwd} refers to the width of the border in pixels.
This also differs from the base functions by directly changing pixels
in the image, in contrast to changing the image on the plot device. 

This region of interest can be specified with arguments or interactively. 
Options allow specifying the rectangle either by
the center or corner(s) as describe below. Options 1 through 5 below require
no interaction with the user and will display the revised image
only if \code{show = TRUE}. Option 6 below requires interaction
with the user. In all cases, the selected coordinates are adjusted to conform
to the dimensions of the image, which can override values in \code{w} or 
\code{h}.

\enumerate{
  \item{\strong{Two Points}.} If values are provided for each of \code{x,y}
    and \code{x2,y2}, these are treated as opposite corners of
    the rectangular selection.
  \item{\strong{Object with \code{loc} slot}.} If \code{x} is an object
    of class \code{Roi}, the coordinates in the \code{loc} slot
    will be extracted and applied to the \code{img} argument.
  \item{\strong{List}.} If \code{x} is a \code{list} of length 2, it
    must hold opposite corners of the rectangular selection. This
    could be the \code{loc} slot from a \code{Roi} object or the result
    of a call to \code{locator(2)}.
  \item{\strong{No Points} (use entire image).} If all of
    \code{x,y, x2,y2, w,h} are missing, a border will be generated for
    the entire image.
  \item{\strong{One Point} (with width and height).} A single value
    can be provided for each of \code{x,y} with named values for
    for \code{w,h} as the width and height of the rectangular selection.
    The point \code{x,y} is interpreted as \emph{either} the center
    (\code{asCorner = FALSE}) \emph{or} the corner of the selection
    (\code{asCorner = TRUE}). If \code{asCorner = TRUE}, the position of the
    corner is determined by the argument \code{which.corner} which can be 
    one of \code{"bottomleft", "topleft", "topright",} or \code{"bottomright"}.
  \item{\strong{Only \code{width} and \code{height}}.} 
    If only \code{(w,h)} are provided
    as named arguments, \code{\link{locator}} will be used to interact
    with the user to identify the point needed to define the rectangular
    selection. The selected point is interpreted as \emph{either} the center
    (\code{asCorner = FALSE}) \emph{or} the corner of the selection
    (\code{asCorner = TRUE}) as described above.
}
}
\examples{

# Example of adding arbitrary 90 pixel x 90 pixel "roi" highlights to one image
  img <- readImage(system.file("extdata", "lighthouse.jpg", package="EBImageExtra"))
  xc <- sample(100:(dim(img)[1] - 100), 8)
  yc <- sample(100:(dim(img)[2] - 100), 8)
  for (i in 1:8) img <- drawROI(img, xc[i], yc[i], w = 90, h = 90)
  plot(img)

# Example of adding a frame to the image itself
  drawROI(img, as.Roi(img), col = "red", lwd = 12, show = TRUE) 

}
\seealso{
\code{\link{getROI}} to get a region of interest from an image;
\code{\link{putROI}} to place an ROI with scaling;
\code{\link{insertROI}} as a convenience function that
  combines calls to \code{getROI}, \code{putROI}
  and \code{drawROI} to place a framed inset in an image.
}
