% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getROI.R
\name{getROI}
\alias{getROI}
\title{Get Region of Interest}
\usage{
getROI(img, x, y, x2, y2, w, h, show, asCorner = FALSE,
  which.corner = c("bottomleft", "topleft", "bottomright", "topright"),
  pch = 3, col = "magenta", border = col, lwd = 2)
}
\arguments{
\item{img}{An \code{Image} object}

\item{x, y}{\code{x,y} coordinates of one or both corners of
rectangular selection \emph{or} the center of the
rectangular selection \emph{or} a list of
corners of the selection (see details in \strong{Selecting
the ROI} below)}

\item{x2, y2}{optional second pair of x and y coordinates when
needed to specify the other corner of the rectangular selection}

\item{w, h}{optional width and height of the rectangular selection
in pixels; required if \code{x2,y2} are missing}

\item{show}{\code{logical} value to redraw the image the selection
outlined by \code{\link{drawROI}}; if \code{missing} or \code{NULL}
\code{show} will set to \code{TRUE} if the user interacts 
with the image and \code{FALSE} if no interaction is required}

\item{asCorner}{\code{logical} value to use the
point \code{x,y} as the corner of the selection or as the center of
the selection}

\item{which.corner}{identifies the
corner of the rectangle specified by \code{x,y};
applies only if \code{asCorner = TRUE}}

\item{pch}{plotting character used by \code{link{locator}} to
indicate mouse clicks, the default value of 3 shows a small cross}

\item{col}{color for plotting character used by \code{link{locator}};
use \code{NA} for no plotting character}

\item{border}{border color of rectangle if \code{show = TRUE}; if not
specified, the value for \code{col} is used}

\item{lwd}{line width of rectangle if \code{show = TRUE}}
}
\value{
The region of interest as an \code{Image} with the
added class of "\code{Roi}" and the slot "\code{loc}"
holding the location of the region of interest.
}
\description{
Get a rectangular region of interest from an image.
}
\section{Selecting the ROI}{


A rectangular region of interest (ROI) can be selected programmatically
or interactively. The ROI is defined by a pair of points (in pixels)
that define a diagonal of the region of interest. The pair of points
can be specified by several means. The function will accept four values
\code{x,y}, and \code{x2,y2} or a list of the two points. Without these
arguments, the function invokes \code{\link{locator}} to allow the
user to define the ROI. The ROI will be trimmed to the
dimensions allowed by the original image. The rectangle can be specified
either by the center or corner(s) as describe below. The first
three options require no interaction with the user and only produce an
image if \code{show = TRUE}. Options 4 and 5 below require interaction
with the user and produce an image if \code{show} is missing or if
\code{show = TRUE}. The returned object is an \code{Image} with
an additional \code{class} and \code{slot} as described in the below.

\enumerate{
  \item{\strong{List}.} If \code{x} is a \code{list} of length 2, it
    is assumed to hold opposite corners of the rectangular selection. This
    is the typical result of a call to \code{locator(2)} after having
    plotted the image.
  \item{\strong{Two Points}.} If values are provided for each of \code{x,y} and
    \code{x2,y2}, these are treated as opposite corners of
    the rectangular selection.
  \item{\strong{One Point} (with width and height).} A single value
    can be provided for each of \code{x,y} with named values for
    for \code{w,h} as the width and height of the rectangular selection.
    The point \code{x,y} is interpreted as \emph{either} the center
    (\code{asCorner = FALSE}) \emph{or} the corner of the selection
    (\code{asCorner = TRUE}). If \code{asCorner = TRUE}, the position of the
    corner is determined by the argument \code{which.corner} which can be 
    one of \code{"bottomleft", "topleft", "topright",} or \code{"bottomright"}.
  \item{\strong{Only \code{width} and \code{height}}.} 
    If only \code{(w,h)} are provided
    as named arguments, \code{\link{locator}} will be used to interact
    with the user to identify the point needed to define the rectangular
    selection. The selected point is interpreted as \emph{either} the center
    (\code{asCorner = FALSE}) \emph{or} the corner of the selection
    (\code{asCorner = TRUE}) as described above.
  \item{\strong{No Points} (choose opposite corners).} If all of \code{x,y, x2,y2, w,h}
    are missing, \code{\link{locator}} will be called to let the user to
    select two points that define opposite corners of the rectangular selection.
}
}

\section{Class "Roi" and "loc" slot}{


\code{EBImage} uses the \code{Image} class to store and process images.
A region of interest is an \code{Image} object with the
additional class named "\code{Roi}" and an additional attribute using
the \code{slot} property of \code{S4} objects. The new \code{slot} or
attribute named "\code{loc}"
holds the location of the region of interest as \code{loc =
list(x = c(x,x2), y = c(y,y2))}.  The attribute \code{"loc"} can be
used to extract the equivalent region of interest from a related image
(\code{img2}) in the following manner. 
\preformatted{
 x <- getROI(img1)
 pp <- attr(x, "loc") ## or pp <- x@loc
 y <- getROI(img2, pp)
}
}

\examples{
# Image from EBImage package
  birds <- readImage(system.file("images", "sample-color.png", package="EBImage"))

# Example specifying one point for center with fixed width and height
  roi1 <- getROI(birds, 160, 255, w = 200, h = 240)
  print(roi1@loc) # one way extract 'loc'
  roi2 <- getROI(birds, 480, 200, w = 200, h = 240)
  print(attr(roi2, "loc")) # an alternative way to extract 'loc'

# Show insets as a combined image
  plotStack(combine(roi1, roi2))

}
\seealso{
\code{\link{putROI}} to place an ROI with scaling and
\code{\link{drawROI}} to draw a frame \emph{within} an image.
}
