% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelStack.R
\name{labelStack}
\alias{labelStack}
\title{Label frames of an image stack}
\usage{
labelStack(x, labels = NULL, nx, col = "white", offset = c(0.05,
  0.05), adj = c(0, 1), ...)
}
\arguments{
\item{x}{A recently plotted \code{Image} object or \code{array} of such
objects}

\item{labels}{Labels to be added to each frame of the image as a
\code{character} vector, or an \code{expression}. If \code{NULL}, the number
of each frame will serve as the label}

\item{nx}{An \code{integer} of length 1 specifying the number of frames in
the x-direction of the image
stack. If missing, a value will be chosen to create an equal number of
frames in the x- and y-directions}

\item{col}{Color of the label to be added}

\item{offset}{A numeric vector of length 1 or 2 specifying the relative position
of the label in each frame where the default values of \code{c(0.05, 0.05)}
place the label at the top left corner}

\item{adj}{One or two values in [0, 1] which specify the \code{x} (and
optionally \code{y}) adjustment of the labels. See \code{\link{text}} for
details.}

\item{...}{Additional values passed to \code{\link{text}} for labeling}
}
\value{
This function is called to label a plotted image. No value is returned.
}
\description{
Add labels to a plotted image stack
}
\details{
Each frame of an image plotted as a raster object will be labeled as
specified by \code{offset} and \code{adj}.
}
\examples{
x <- readImage(system.file("images", "sample-color.png", package="EBImage"))
x <- EBImage::untile(x, c(4, 3), lwd = 0)
idx <- sample(1:12)
plotStack(x[,,,idx], label = paste("Frame", 1:12))
labelStack(x, labels = idx, col = "yellow", offset = 0.5, adj = 0.5, cex = 2)

}
\seealso{
\code{\link{plotStack}}, \code{\link{locatorStack}}
}
