% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelStack.R
\name{labelStack}
\alias{labelStack}
\title{Label Frames of an Image Stack}
\usage{
labelStack(x, labels = TRUE, nx, col = "white", offset = c(0.05,
  0.05), adj = c(0, 1), ...)
}
\arguments{
\item{x}{A recently plotted \code{Image} object or \code{array} of such
objects}

\item{labels}{Either a \code{logical} value (default of \code{TRUE})
to number each frame \emph{or} a vector of labels to be applied to each frame}

\item{nx}{The number of frames in the x-direction of the image stack.
If missing, a square tile of images will be assumed}

\item{col}{Color of the label to be added}

\item{offset}{A numeric vector of length 1 or 2 specifying the position
of the label in each frame where the default values of \code{c(0.05, 0.05)}
place the label at the top left corner}

\item{adj}{One or two values in [0, 1] which specify the \code{x} (and
optionally \code{y}) adjustment of the labels. See the \code{\link{text}}
function for details.}

\item{...}{Additional values passed to \code{\link{text}} for labeling}
}
\value{
The image \code{x} is invisibly returned to permit passing the argument through
a pipe. This function primarily called to label a plotted image.
}
\description{
Add labels to a plotted image stack
}
\details{
This code is called after having plotted an \code{Image} object, typically
by \code{\link{plotStack}} or \code{plot(img, all = TRUE)}. If \code{labels}
is \code{TRUE}, each frame of the image will be labeled by frame number.
If \code{FALSE}, no labelling will occur. Otherwise, labels generated by 
\code{as.graphicsAnnot} will be placed in each frame according to
\code{offset} and \code{adj}.
}
\examples{
lighthouse <- readImage(system.file("extdata", "lighthouse.jpg", package="EBImageExtra"))
x <- EBImage::untile(lighthouse, c(4, 3), lwd = 0)
idx <- sample(1:12)
plotStack(x[,,,idx], label = TRUE)
labelStack(x, labels = paste("Original:", idx), offset = 0.5, adj = 0.5, col = "yellow")

}
\seealso{
\code{\link{plotStack}}, which allows for automatically labeling the plotted
image and \code{\link{locatorStack}}, that provides an interface for
selecting frames from the stack.
}
