% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locatorStack.R
\name{locatorStack}
\alias{locatorStack}
\title{Select Frames from a Plotted Image Stack}
\usage{
locatorStack(x, labels, nx, col = "red", offset = c(0.05, 0.05),
  adj = c(0, 1), value = c("frames", "image"), ...)
}
\arguments{
\item{x}{An \code{Image} object or \code{array} of such objects}

\item{labels}{Optional labels to be added to each selected frame. If
missing, the number of each frame will serve as the label}

\item{nx}{The number of frames in the x-direction of the image stack.
If missing, a square tile of images will be assumed}

\item{col}{Color of the label to be added}

\item{offset}{A numeric vector of length 2 specifying the relative position
of the label in each frame}

\item{adj}{One or two values in [0, 1] which specify the \code{x} (and
  optionally \code{y}) adjustment of the labels. See the function
\code{\link{text}} for further details}

\item{value}{Character vector indicating whether to return a numeric vector
of found frames or a new \code{Image} (see Details)}

\item{...}{Additional values passed to \code{\link{text}} for labeling}
}
\value{
\strong{Either} a numeric vector of the selected frames \strong{or} an
\code{Image} object of the selected frames in the order that they
were selected.
}
\description{
Identify and select frames in a plotted image stack.
}
\details{
The \code{base} function \code{\link{locator}} will be used to interact
with the user to identify frames in the plotted
image stack. Frames selected on the image will be labeled and a count
of the selected (unique) frames will be reported in a running tally on the 
console. Selection is stopped by pressing any mouse button other than the
primary button or by pressing the \code{ESC} key. The selected frames, in
the order that they were selected, will be returned as a vector of integers
if \code{value} is \code{"frames"} or as a new \code{Image} object of the
selected frames if \code{value} is \code{"image"}.
}
\seealso{
\code{\link{plotStack}}, which allows for automatically labeling the plotted
image and \code{\link{labelStack}}, that places custom labels on each 
frame of the image stack.
}
