% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locatorStack.R
\name{locatorStack}
\alias{locatorStack}
\title{Interact with plotted image stack to select frames}
\usage{
locatorStack(x, labs, nx, col = "red", offset = c(0.05, 0.05),
  adj = c(0, 1), ...)
}
\arguments{
\item{x}{An \code{Image} object of \code{array} of such objects}

\item{labs}{Optional labels to be added to each found frame. If
\code{missing}, the number of each frame will serve as the label}

\item{nx}{An \code{integer} of length 1 specifying the width of the image
stack. If missing, a value will be chosen to create an equal number of
frames in the x- and y-directions}

\item{col}{Color of the label to be added}

\item{offset}{A numeric vector fo length 2 specifying the relative position
of the label in each frame where the default values of \code{c(0.05, 0.05)}
place the label at the top left corner}

\item{adj}{One or two values in [0, 1] which specify the \code{x} (and
optionally \code{y}) adjustment of the labels. See \code{\link{text}} for
details.}

\item{...}{Additional values passed to \code{\link{text}} for labeling}
}
\value{
A vector of selected frames
}
\description{
The function \code{\link{locator}} is called to identify and select frames
on an image stack.
}
\details{
The base function \code{\link{locator}} will be called until canceled by the
user to identify frames in the *previously* plotted image stack. The
selected frame will be labeled and a count of the (unique) selected frames
will be reported in a running tally. The selected frames, in the order that
they were selected, will be returned as a vector of integers.
}
\seealso{
\code{\link{labelStack}}, \code{\link{plotStack}}
}
