% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locatorStack.R
\name{locatorStack}
\alias{locatorStack}
\title{Interact with plotted image stack to select frames}
\usage{
locatorStack(x, labels, nx, col = "red", offset = c(0.05, 0.05),
  adj = c(0, 1), value = c("frames", "image"), ...)
}
\arguments{
\item{x}{An \code{Image} object of \code{array} of such objects}

\item{labels}{Optional labels to be added to each found frame. If
\code{missing}, the number of each frame will serve as the label}

\item{nx}{An \code{integer} of length 1 specifying the width of the image
stack. If missing, a value will be chosen to create an equal number of
frames in the x- and y-directions}

\item{col}{Color of the label to be added}

\item{offset}{A numeric vector fo length 2 specifying the relative position
of the label in each frame where the default values of \code{c(0.05, 0.05)}
place the label at the top left corner}

\item{adj}{One or two values in [0, 1] which specify the \code{x} (and
optionally \code{y}) adjustment of the labels. See \code{\link{text}} for
details.}

\item{value}{Character vector indicating the return value. The default of
\code{"found"} returns the selected frames. The option \code{"image"}
returns an \code{Image} or \code{Image} array composed of the selected frames.}

\item{...}{Additional values passed to \code{\link{text}} for labeling}
}
\value{
\strong{Either} a numeric vector of the selected frames \strong{or} an
\code{Image} object of the selected frames in the order that they
were selected.
}
\description{
The function \code{\link{locator}} is called to identify and select frames
on an image stack.
}
\details{
The \code{base} function \code{\link{locator}} will be called until
stopped by the user to identify frames in the \emph{previously} plotted
image stack. The selected frame on the image will be labeled and a count
of the selected (unique) frames will be reported in a running tally on the 
console. The selected frames, in the order that they were selected, will
either be returned as a vector of integers if \code{value} is "frames", or
if \code{value} is "image", a new \code{Image} object of the selected
frames will be returned.
}
\seealso{
\code{\link{labelStack}}, \code{\link{plotStack}}
}
