% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locatorStack.R
\name{locatorStack}
\alias{locatorStack}
\title{Select Frames from a Plotted Image Stack}
\usage{
locatorStack(x, labels, nx, col = "red", asImage = FALSE,
  offset = c(0.05, 0.05), adj = c(0, 1), combine = TRUE, ...)
}
\arguments{
\item{x}{An \code{Image} object or \code{array} of such objects}

\item{labels}{Optional labels to be added to each selected frame. If
missing, the number of each frame will serve as the label otherwise
the argument will be adjusted with \code{as.graphicsAnnot}}

\item{nx}{The number of frames in the x-direction of the image stack.
If missing, a square tile of images will be assumed}

\item{col}{Color of the label to be added}

\item{asImage}{\code{logical} value (default of \code{FALSE}) indicating
whether the function returns the selection as \code{Image}(s) or
as a numeric vector of found frames (see Details)}

\item{offset}{A numeric vector of length 2 specifying the relative position
of the label in each frame}

\item{adj}{One or two values in [0, 1] which specify the \code{x} (and
optionally \code{y}) adjustment of the labels. See the function
\code{\link{text}} for further details}

\item{combine}{\code{logical} value (default of \code{TRUE}) to return
the selected frames combined as a single image rather than a list}

\item{...}{Additional values passed to \code{\link{text}} for labeling}
}
\value{
A numeric vector of the selected frames if \code{asImage = FALSE} or
either a list of selected \code{Images} frames or combined \code{Image}
frames in the order that they were selected.
}
\description{
Identify and select frames in a plotted image stack.
}
\details{
The \code{\link{locator}} function will be called to interactively
to identify frames in the plotted
image stack. Selected frames will be labeled on the image and a count
of the unique selected frames will be reported in a running tally on the 
console. Selection is stopped by pressing any mouse button other than the
primary button or by pressing the \code{ESC} key. The selected frames, in
the order that they were selected, will be returned as a vector of integers
if \code{asImage = FALSE}. If \code{asImage = TRUE}, the selected images
will be returned as a list of separate frames if \code{combine = FALSE}
or as a single image if \code{combine = TRUE}.
}
\seealso{
\code{\link{plotStack}}, which allows for automatically labeling the plotted
image and \code{\link{labelStack}}, that places custom labels on each 
frame of the image stack.
}
