% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni.R
\name{ni}
\alias{ni}
\alias{vni}
\title{Vectorized normalize function}
\usage{
ni(x, ir = c(0, 2^-4), separate = TRUE, ft = c(0, 1))
}
\arguments{
\item{x}{An \code{Image} object or array or a list of \code{Image} objects}

\item{ir}{A numeric vector or list vectors of 2 values to set the range of
input intensity. This value is passed to \code{inputRange} for
\code{\link[EBImage]{normalize}}}

\item{separate}{If \code{TRUE} or \code{ir} is \code{NULL}, normalize each
frame separate}

\item{ft}{A numeric vector or \code{list} of vectors of 2 values as the
target minimum and maximum intensity values after normalization}
}
\value{
An \code{Image} object or array or a \code{list} of such objects containing
the transformed object(s).
}
\description{
Vectorized wrapper for \code{\link[EBImage]{normalize}} with a default
\code{inputRange} parameter appropriate for 12-bit images.
}
\details{
See \code{\link[EBImage]{normalize}} for details. The default here is to use
an \code{inputRange} of \code{c(0,2^-4)}, which is appropriate for
12-bit images. This function provides an easier-to-type replacement for 
\code{\link{normalize}} while extending it to accept lists of images.
}
