% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStack.R
\name{plotStack}
\alias{plotStack}
\title{Plot image stack with optional labels}
\usage{
plotStack(x, labels = FALSE, nx, ...)
}
\arguments{
\item{x}{An \code{Image} object of \code{array} of such objects}

\item{labels}{A \code{logical} value, a \code{character} vector, or an
\code{expression} specifying the addition of labels. No labels are added with
the default value of \code{FALSE}}

\item{nx}{An \code{integer} of length 1 specifying the width of the image
stack. If missing, a value will be chosen to create an equal number of
frames in the x- and y-directions}

\item{...}{Additional parameters passed to \code{\link{labelStack}}}
}
\value{
No value is returned. This function is called to plot an entire \code{Image}
object.
}
\description{
This is a wrapper to plot all frames of an image as a raster object with the
option of labeling each frame.
}
\details{
The image in \code{x} will be plotted by the command
\code{plot(x, all = TRUE, nx = nx)}. If \code{labels} is not \code{FALSE},
labels will be generated and added with \code{\link{labelStack}} where
\code{labels = TRUE} will simply number each frame. Other values of
\code{labels} will be used as a label, replicating the values in
\code{labels} as necessary.
}
\examples{
x <- readImage(system.file("images", "sample-color.png", package="EBImage"))
x <- EBImage::untile(x, c(4, 3), lwd = 0)
plotStack(x, labels = paste("Frame", 1:12))

}
\seealso{
\code{\link{labelStack}}, \code{\link{locatorStack}}
}
