% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStack.R
\name{plotStack}
\alias{plotStack}
\title{Plot Image Stack with Optional Labels}
\usage{
plotStack(x, labels = FALSE, nx, cex = 1, ...)
}
\arguments{
\item{x}{An \code{Image} object or \code{array} of such objects}

\item{labels}{If \code{FALSE}, no labels are added. If \code{TRUE} or
\code{NULL}, frames will be sequentially number. Other values of
\code{labels} will be used to label for each frame}

\item{nx}{The number of frames in the x-direction of the image stack.
If missing, a square tile of images will be assumed}

\item{cex}{The character expansion for \code{labels}}

\item{...}{Additional parameters passed to \code{\link{plot.Image}}}
}
\value{
The image \code{x} is invisibly returned to permit passing the argument
through a pipe. This function primarily called to plot an entire
\code{Image} object.
}
\description{
This is a wrapper to plot all frames of an image as a raster object with the
option of labeling each frame.
}
\details{
The image in \code{x} will be plotted by the command
\code{plot(x, all = TRUE, nx = nx)}. If \code{labels} is not \code{FALSE},
labels will be generated and added with \code{\link{labelStack}} where
\code{labels = TRUE} will simply number each frame. Other values of
\code{labels} will be used as a label, replicating the values in
\code{labels} as necessary.
}
\examples{
# Sample color image
  img <- readImage(system.file("inst", "extdata", "lighthouse.jpg", package="EBImageExtra"))

# separate into 12 smaller images (frames)
  x <- EBImage::untile(img, c(4, 3), lwd = 0)
  plotStack(x, labels = paste("Frame", 1:12))

}
\seealso{
\code{\link{labelStack}}, to place custom labels on each frame of the
image stack and \code{\link{locatorStack}}, that provides an interface for
selecting frames from the stack.
}
