% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStack.R
\name{plotStack}
\alias{plotStack}
\title{Plot image stack with optional labels}
\usage{
plotStack(x, labels = FALSE, nx, ...)
}
\arguments{
\item{x}{An \code{Image} object or \code{array} of such objects}

\item{labels}{If \code{FALSE}, no labels are added. If \code{TRUE} or
\code{NULL}, frames will be sequentially number. Other values of
\code{labels} will be used to label for each frame}

\item{nx}{The number of frames in the x-direction of the image stack.
If missing, a square tile of images will be assumed}

\item{...}{Additional parameters passed to \code{\link{plot.Image}}}
}
\value{
No value is returned. This function is called to plot an entire \code{Image}
object.
}
\description{
This is a wrapper to plot all frames of an image as a raster object with the
option of labeling each frame.
}
\details{
The image in \code{x} will be plotted by the command
\code{plot(x, all = TRUE, nx = nx)}. If \code{labels} is not \code{FALSE},
labels will be generated and added with \code{\link{labelStack}} where
\code{labels = TRUE} will simply number each frame. Other values of
\code{labels} will be used as a label, replicating the values in
\code{labels} as necessary.
}
\examples{
x <- readImage(system.file("images", "sample-color.png", package="EBImage"))
x <- EBImage::untile(x, c(4, 3), lwd = 0)
plotStack(x, labels = paste("Frame", 1:12))

}
\seealso{
\code{\link{labelStack}}, to place custom labels on each frame of the
image stack and \code{\link{locatorStack}}, that provides an interface for
selecting frames from the stack.
}
