% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleBar.R
\name{scaleBar}
\alias{scaleBar}
\title{Add Scale Bar to Image}
\usage{
scaleBar(
  x,
  y = NULL,
  width,
  distance,
  label = NULL,
  col = "white",
  col.text,
  cex = 3/4,
  adj = c(0.5, -0.5),
  col.line,
  lwd = 1,
  lend = 1,
  ...
)
}
\arguments{
\item{x, y}{Central coordinates of the scale bar in pixels}

\item{width}{Width of the scale bar in pixels}

\item{distance}{Distance represented by the scale bar in microns}

\item{label}{A suitable label that can be coerced to \code{character} or
an \code{expression}. If \code{NULL}, a label will be generated from
the \code{distance} argument. To exclude any label, set this argument to
\code{""} or to \code{NA}}

\item{col}{Default color (white) for \code{col.text} and \code{col.line}}

\item{col.text, cex, adj, ...}{Parameters passed to the \code{\link{text}}
function used to add the label}

\item{col.line, lwd, lend}{Parameters passed to the \code{\link{lines}}
function used to draw the horizontal scale bar}
}
\value{
This function is called for side effects. No value is returned.
}
\description{
Place a single horizontal scale bar \code{width} pixels wide corresponding
to a physical distance \code{distance} with an optional \code{label}.
}
\details{
This is a utility to place a labeled scale bar on a raster image. 
A default label in microns will
be added if \code{label} is \code{NULL}. The label is centered at the
position specified by \code{x,y} according to the values in \code{adj}.
The default value for \code{adj} of \code{c(0.5,-0.5)} centers the label
and places it 0.5 character widths above \code{x,y}.
 
The \code{x,y} coordinates can be passed in a plotting structure
appropriate for \code{\link{xy.coords}}. This function attempts to parse
\code{distance} and \code{width} from the remaining arguments if they are
not named and if \code{x} is a list with components named "x" and "y". This 
allows one to interactively place a scale bar with a function
call such as in the following example. In this example, \code{ppm} stands
for "pixels per micrometer" and is the scale factor appropriate for the
given micrograph.
This call will place a labeled 25 micron scale bar on the image.  
\preformatted{
> scaleBar(locator(), 25*ppm, 25)
}

Note that this is \strong{not} a vectorized function. Only the last
value of \code{(x, y)} will be used to place the scale bar. This allows
multiple clicks to \code{locator()} to get the position just right
with only the last click being used.
}
