% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobel.R
\name{sobel}
\alias{sobel}
\title{Sobel filter for edge detection}
\usage{
sobel(x)
}
\arguments{
\item{x}{An \code{Image} object or array}
}
\value{
An image object of the same size with edges highlighted.
}
\description{
Detect edges in an \code{Image} object
}
\details{
The following matrix and its transpose are sequentially applied in
\code{\link{filter2}} to highlight edges in the image.
\preformatted{
     [,1] [,2] [,3]
[1,]    1    0   -1
[2,]    2    0   -2
[3,]    1    0   -1
}
}
