% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackObjects2.R
\name{stackObjects2}
\alias{stackObjects2}
\title{Place Detected Objects with Bounding Box in an Image Stack}
\usage{
stackObjects2(mask, ref, combine = TRUE, bg.col = "black", ext)
}
\arguments{
\item{mask}{An \code{Image} object or an \code{array} containing object
masks. Object masks are sets of pixels with the same unique integer value.}

\item{combine}{If \code{mask} contains multiple images, should the resulting
\code{list} of image stacks be combined using \code{\link[EBImage]{combine}}
into a single image stack.}

\item{bg.col}{Background pixel color.}

\item{ext}{A \code{numeric} controlling the size of the output image. If
missing, \code{ext} is estimated from data as described in
\code{\link[EBImage]{stackObjects}}.}
}
\value{
An \code{Image} object containing the stacked objects contained in
\code{mask}. If \code{mask} contains multiple images and if \code{combine}
is \code{TRUE}, \code{stackObjects2} returns a single combined image
otherwise a \code{list} of \code{Image} objects is returned.
}
\description{
Place objects in an image stack with the same logic as
\code{\link[EBImage]{stackObjects}} except the entire bounding box is
included for each object in the image stack.
}
\details{
Like \code{stackObjects}, \code{stackObjects2} creates a set of \code{n}
images of size \code{c(2*ext+1,2*ext+1)}, where \code{n} is the number of
objects in \code{mask}, and places each object of \code{mask} in this set.

If \code{ext} is missing, it is estimated with the 98\% quantile of the
\code{m.majoraxis/2} extracted from \code{computeFeatures.moment}.
}
