% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usm.R
\name{usm}
\alias{usm}
\title{Unsharp mask}
\usage{
usm(x, radius = 50, amount = 0.2, inputRange)
}
\arguments{
\item{x}{An \code{Image} object or array of \code{Image} objects}

\item{radius}{\code{Integer} value to determine Gaussian blur effect, 
default value of 50}

\item{amount}{A \code{numeric} value between 0 and 1 specifying the amount of
blur to apply, default value of 0.2.}

\item{inputRange}{value A optional \code{numeric} vector of 2 values handed to
\code{\link[EBImage]{normalize}} to normalize the final image. If missing,
the range of the input object will be used}
}
\value{
A sharpened \code{Image} object of the same size
}
\description{
Sharpen image with unsharp mask algorithm
}
\details{
The unsharp mask algorithm is applied.
}
\examples{
x <- readImage(system.file("images", "sample-color.png", package="EBImage"))
x <- x[156:265, 166:300,]
y <- usm(x, radius = 15, amount = 0.8)
plotStack(combine(x, y), label = c("Original", "Unsharp mask"), cex = 2)

}
